+++ contracts/mintAllocated.sol
@@ -72,7 +72,7 @@ contract MintAllocated is
         isMinter(minter)
     {
         require(
-            type(uint).max - mintAllocation[minter] >= amount,
+            type(uint256).max - mintAllocation[minter] >= amount,
             ErrorCoded.ERR_ARITHMETIC_OVERFLOW_MINT
         );
         unchecked {
